+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `awless whoami` now returns your identity, your attached (i.e. managed), inlined and group policies
 - Rudimentary security groups port scanner inspector via `awless inspect -i port_scanner`
 - Template: compile time check of undefined or unused references
+- Run official remote templates without specifying full url: `awless run repo:create_vpc`
 - [#78](https://github.com/wallix/awless/issues/78): Show progress when uploadgin object to storage
 - [#81](https://github.com/wallix/awless/issues/81): Global force flag `--force` to bypass confirm prompt